Added the application file
Create an enumeration with a checkbox that stores the names of 8 disciplines of professional training—4 disciplines of the bachelor's and master's degrees. Print the list of disciplines using constants (number and name). Give the user the opportunity to enter the number of disciplines he would like to study in the form of a list, separating the numbers with commas. Print the name of the author of the program and lists of disciplines separately for each educational level.